requirements.txt
asgiref==3.3.4
Django==3.0.1
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-js-asset==1.2.2
django-material-admin==1.7.8
django-robots==4.0
django-taggit==1.5.1
PyMySQL==1.0.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==4.1.1
Pillow==8.2.0